% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropEmptyDates.R
\name{dropEmptyDates}
\alias{dropEmptyDates}
\title{Drops empty dates columns}
\usage{
dropEmptyDates(st)
}
\arguments{
\item{st}{data.frame}
}
\value{
data.frame without half empty date columns
}
\description{
Drops half empty columns dates.
}
\details{
Any column which the name can be transformed to
a date with as.Date(format="%Y-%m-%d") will be
evaluated, if more then half or equal of the values of
the date column is NA, it will be droped. It will return
a data.frame and wont modify the original.
}
\keyword{dates}
\keyword{format}
