% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundfiDoAll.R
\name{fundfiDoAll}
\alias{fundfiDoAll}
\title{Create finantial reports from tickers}
\usage{
fundfiDoAll(
  tickers,
  type = "10-Q",
  count = 1,
  untile = NULL,
  drop.empty.dates = FALSE,
  keep.xbrl = FALSE,
  statements.dir = FALSE,
  force.new = FALSE,
  xbrl.cache = "xbrl.Cache",
  xbrl.rds.cache = "xbrl.rds.Cache"
)
}
\arguments{
\item{tickers}{A ticker character value or vector}

\item{type}{Type of filling like 10-Q or 10-K}

\item{count}{Number of maximum filings. If until parameter is set, the
number of filings returned could be less.}

\item{drop.empty.dates}{Defaults to FALSE. Bolean that indicates
if it should keep or drop half empty date columns in statements.}

\item{keep.xbrl}{Defaults to FALSE. If TRUE, the list of XBRL will also
be returned as an element of the list returned called original.xbrl .}

\item{statements.dir}{Defaults to NULL. If a string is passed,
each ticker list of statement will be saved as a .rds in the
folder with the name specified.}

\item{force.new}{Defaults to FALSE. If TRUE, ignores XBRL rds cache.}

\item{xbrl.cache}{Folder where XBRL taxonomies are saved.
If null then it wont save cache.}

\item{xbrl.rds.cache}{Folder where XBRL lists are saved as .rds
If null, it wont save cache.}

\item{until}{yyyymmdd fillings are downloaded starting from the until
parameter until the oldest, or until it reaches the count limit.}
}
\value{
List that contains the balance sheet, income statement and
cash flow of each ticker. The names of the list are the tickers.
}
\description{
Fetches the XBRLs of the tickers specified from the SEC's EDGAR,
and creates finantial reports.
}
\details{
This function will fetch the XBRL documents of
the specified tickers, download them and format them
to create balance sheet, income statement and cash flow.
It will return a list of lists, each on of them
corresponding to the ticker name and as his elements,
data.frames containing the finantial report.
There is also the option to save each ticker list
as a .rds in a folder using the statements.dir
parameter. For network performance, XBRL taxonomies
and XBRL list are cached but this behaviour can
be modified.
}
\examples{
getXBRL(c("AAPL", "TSLA", "GE"), type = "10-Q")
}
