% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFullDataDates.R
\name{getFullDataDates}
\alias{getFullDataDates}
\title{Eliminates empty dates columns}
\usage{
getFullDataDates(st)
}
\arguments{
\item{st}{data.frame returned by getStatement}
}
\value{
data.frame with dates columns that are half full of data
}
\description{
Eliminates dates columns that are half empty
}
\keyword{dates}
\keyword{format}
