% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getXBRL.R
\name{getXBRL}
\alias{getXBRL}
\title{Fetch a list of XBRLs}
\usage{
getXBRL(
  tickers = NULL,
  type = "10-Q",
  count = 1,
  until = NULL,
  force.new = FALSE,
  xbrl.cache = "xbrl.Cache",
  xbrl.rds.cache = "xbrl.rds.Cache"
)
}
\arguments{
\item{tickers}{A ticker character value or vector}

\item{type}{Defaults to 10-Q. Type of filling like 10-Q or 10-K}

\item{count}{Number of maximum filings. If until parameter is set, the
number of filings returned could be less.}

\item{until}{yyyymmdd fillings are downloaded starting from the until
parameter until the oldest, or until it reaches the count limit.}

\item{force.new}{Defaults to FALSE. If TRUE.}

\item{xbrl.cache}{Folder where XBRL taxonomies are saved.
If null then it wont save cache.}

\item{xbrl.rds.cache}{Folder where XBRL lists are saved as .rds
If null, it wont save cache.}
}
\value{
List with ticker names which have lists of data.frames, formated
as XBRL
}
\description{
Fetches the XBRLs of the tickers specified from the SEC's EDGAR.
}
\details{
This function will fetch the XBRL documents of
the specified tickers. It will return a list of
XBRL list. Each XBRL list contains the data.frames
returned by the library XBRL. Each ticker XBRL from
a data are cached as a .rds in the folder xbrl.rds.Cache.
All XBRL taxonomies are cached in the folder xbrl.Cache.
These behaviors can be modified using parameters.
If any ticker doesn´t exist, it will be dropped. It´s
recomended to check if the length of the tickers
is equal to the returned list.
}
\examples{
getXBRL(c("AAPL", "TSLA", "GE"), type = "10-Q")
}
\keyword{XBRL}
