% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStatement.R
\name{getStatement}
\alias{getStatement}
\title{Create finantial reports}
\usage{
getStatement(xbrl.vars = NULL, statement = "balance_sheet",
  custom.roleID = NULL)
}
\arguments{
\item{xbrl.vars}{XBRL list of data.frames}

\item{statement}{Type of statement. Can be balance_sheet, income_statement and cash_flow}
}
\value{
A data.frame as a financial report
}
\description{
This function will create finantial reports of a list
of data.frames with the XBRL format.
}
\examples{
getStatement(getXBRL(c("PG", type="10-Q"))$PG, statement = income_statement)
}
\keyword{Finantial}
\keyword{Statement}
