% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStatement.R
\name{getStatement}
\alias{getStatement}
\title{Create finantial report}
\usage{
getStatement(xbrl.vars = NULL, statement = "balance_sheet",
  custom.roleID = NULL)
}
\arguments{
\item{xbrl.vars}{XBRL list of data.frames}

\item{statement}{Type of statement. Can be balance_sheet, income_statement and cash_flow}

\item{custom.roleId}{Character to fetch roleId using description}
}
\value{
data.frame as a financial report
}
\description{
Creates a data.frame containing a finantial
report using a XBRL list of data.frame.
}
\examples{
getStatement(getXBRL(c("PG", type="10-Q"))$PG, statement = "income_statement")
}
\keyword{Finantial}
\keyword{Statement}
