% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getXBRL.R
\name{getXBRL}
\alias{getXBRL}
\title{Get a list of XBRL}
\usage{
getXBRL(tickers = NULL, type = "10-Q", new = FALSE)
}
\arguments{
\item{tickers}{A ticker character value or vector}
}
\value{
List with ticker names which have lists of data.frames, formated as XBRL
}
\description{
This function will fetch the XBRL documents of
the specified tickers. It will return a list of
XBRL list.
}
\examples{
getXBRL(c("AAPL", "TSLA", "GE"), type = "10-Q")
}
\keyword{XBRL}
