% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printStatement.R
\name{printStatement}
\alias{printStatement}
\title{Print financial statement}
\usage{
printStatement(statement)
}
\arguments{
\item{statement}{data.frame returned by getStatement}
}
\value{
Prints to console or to .md the financial statement
}
\description{
This function prints a financial statement returned by
getStatement.
}
\examples{
printStatement(getStatement(allXBRLs$PG, statement = "income_statement"))
}
\keyword{Print}
\keyword{financial}
\keyword{statement}
